mcdonalds_app/serializers.py
from rest_framework import serializers

from mcdonalds_app.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = "__all__"
