requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
django-debug-toolbar==3.2.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
h11==0.14.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PySocks==1.7.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
